Using EXTRA_VERIFY_ARGS env var to skip -Werror.
TinyXML has a warning that becomes an error when Arduino compiles
with -Werror. The `platform.local.txt` file does not work to
override this setting, so supplying it directly when invoking
the arduino app.

Support for EXTRA_VERIFY_ARGS was added in a branch to
Adafruit's travis-ci-arduino project, so using that branch for
CI until pull request accepted.

Also using EXTRA_VERIFY_ARGS when verifying in the Makefile
for local builds.